raspi/requirements.txt
gpiozero
lgpio
python-socketio
websocket-client
hx711